requirements.txt
gradio==4.44.0
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.10.0.84
matplotlib==3.9.1
pytorch-grad-cam==1.5.2
huggingface-hub==0.24.6
python-dotenv==1.0.1
requests==2.32.3
tenacity==9.0.0
pytest==8.3.2

